refactor: split db/src/stats/mod.rs into mod.rs + compute.rs (#1948)
## Summary
- Closes #1932. `db/src/stats/mod.rs` crossed the ~800-line soft
file-length
cap (804 lines). Mechanically split it into `mod.rs`
(cache/error/re-export
scaffolding: `StatsError` + the `BooksError` conversion, the
process-wide
  `Cache`/`STATS_TTL_SECS` machinery, the `user_stats`/`user_stats_at`
entrypoint) and a new `stats/compute.rs` (the SQL-heavy `compute` body
and
its private per-metric query helpers + the
`TOP_N`/`FINISHED_BOOKS_LIMIT`/
  `SESSION_BOOK_SECS`/`SESSION_TIME_SECS`/`FINISHED_EVENTS` constants).
- No behavior change — pure move, no logic rewrites. `mod.rs` is now 122
  lines, `compute.rs` is 725 lines, both under the cap.
- Updated `docs/architecture.md`'s `stats/` entry to describe the new
  `mod.rs`/`compute.rs` split.

## Test plan
- [x] `cargo fmt -p omnibus-db` — PASS (clean)
- [x] `cargo clippy -p omnibus-db --all-targets -- -D warnings` — PASS
- [x] `cargo test -p omnibus-db --lib stats::` — PASS (48/48; every
      `stats/tests.rs`, `stats/book/tests.rs`, `stats/pages.rs` test)
- [x] `cargo test -p omnibus-db` — 2042 passed, 1 failed. The failure

(`audiobook::cover::tests::extract_cover_returns_none_when_valid_audio_has_no_embedded_picture`)
is pre-existing and unrelated to this change: it panics because the
`test_data/audiobooks/public_domain/...` fixture asset isn't present
      in this sandbox (`just fixtures` wasn't run — no nix/just tooling
available here). `db/src/audiobook/cover.rs` is untouched by this PR.
- [x] `wc -l db/src/stats/mod.rs` → 122, `wc -l db/src/stats/compute.rs`
→ 725
      — both under the ~800-line cap

## Version
- [x] **Patch** — default (unlabeled, tech-debt refactor, no behavior
change)

## Notes
- Pure mechanical extraction per the issue's suggested approach; no SQL
or
  logic was changed, only moved.


---
_Generated by [Claude
Code](https://claude.ai/code/session_01DPtZ3DUpjpSdPSEZZaHiYE)_

---------
